+++ database/migrations/2021_06_13_171118_create_apartment_views_table.php
@@ -15,7 +15,7 @@ class CreateApartmentViewsTable extends Migration
     {
         Schema::create('apartment_views', function (Blueprint $table) {
                 $table->id();
-                $table->integer('id_apartment');
+                $table->unsignedBigInteger('id_apartment');
                 $table->timestamps();
         });
     }
